// backend/src/iam/enums/permission.enum.ts
export enum Permission {
  APPOINTMENTS_CREATE = 'appointments:create',
  APPOINTMENTS_READ_SELF = 'appointments:read:self',
  APPOINTMENTS_READ_ASSIGNED = 'appointments:read:assigned',
  APPOINTMENTS_READ_ALL = 'appointments:read:all',
  APPOINTMENTS_UPDATE = 'appointments:update',
  APPOINTMENTS_DELETE = 'appointments:delete',
}